+++ projects/igniteui-angular/migrations/update-11_1_0/index.spec.ts
@@ -252,6 +252,219 @@ export class IconTestComponent {
             </igx-chips-area>`);
     });
 
+    it('should replace IgxTabsComponent event name onTabItemSelected with tabItemSelected', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/tabs.component.html`,
+            `<igx-tabs (onTabItemSelected)="tabSelected()"></igx-tabs>`
+        );
+
+        const tree = await runner
+            .runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/tabs.component.html')
+        ).toEqual(`<igx-tabs (tabItemSelected)="tabSelected()"></igx-tabs>`);
+    });
+
+    it('should replace IgxTabsComponent event name onTabItemDeselected with tabItemDeselected', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/tabs.component.html`,
+            `<igx-tabs (onTabItemDeselected)="tabDeselected()"></igx-tabs>`
+        );
+
+        const tree = await runner
+            .runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/tabs.component.html')
+        ).toEqual(`<igx-tabs (tabItemDeselected)="tabDeselected()"></igx-tabs>`);
+    });
+
+    it('should replace IgxListComponent event name onLeftPan with leftPan', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/list.component.html`,
+            `<igx-list [allowLeftPanning]="true" (onLeftPan)="leftPanPerformed($event)">`
+        );
+
+        const tree = await runner
+            .runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/list.component.html')
+        ).toEqual(`<igx-list [allowLeftPanning]="true" (leftPan)="leftPanPerformed($event)">`);
+    });
+
+    it('should replace IgxListComponent event name onRightPan with rightPan', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/list.component.html`,
+            `<igx-list [allowRightPanning]="true" (onRightPan)="rightPanPerformed($event)">`
+        );
+
+        const tree = await runner
+            .runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/list.component.html')
+        ).toEqual(`<igx-list [allowRightPanning]="true" (rightPan)="rightPanPerformed($event)">`);
+    });
+
+    it('should replace IgxListComponent event name onPanStateChange with panStateChange', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/list.component.html`,
+            `<igx-list (onPanStateChange)="panStateChange($event)"></igx-list>`
+        );
+
+        const tree = await runner
+            .runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/list.component.html')
+        ).toEqual(`<igx-list (panStateChange)="panStateChange($event)"></igx-list>`);
+    });
+
+    it('should replace IgxListComponent event name OnItemClicked with itemClicked', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/list.component.html`,
+            `<igx-list (onItemClicked)="onItemClicked($event)"></igx-list>`
+        );
+
+        const tree = await runner
+            .runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/list.component.html')
+        ).toEqual(`<igx-list (itemClicked)="onItemClicked($event)"></igx-list>`);
+    });
+
+    it('should replace IgxNavbarComponent event name onAction with action', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/navbar.component.html`,
+            `<igx-navbar (onAction)="actionExc($event)" title="Sample App" actionButtonIcon="menu"></igx-navbar>`
+        );
+
+        const tree = await runner
+            .runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/navbar.component.html')
+        ).toEqual(`<igx-navbar (action)="actionExc($event)" title="Sample App" actionButtonIcon="menu"></igx-navbar>`);
+    });
+
+    it('should update Excel exporter onExportEnded event name to exportEnded', async () => {
+        pending('set up tests for migrations through lang service');
+        appTree.create(
+            '/testSrc/appPrefix/component/excel-export.component.ts',
+`import { Component } from '@angular/core';
+import { IgxExcelExporterService } from "igniteui-angular";
+
+@Component({
+    selector: "app-excel-export",
+    styleUrls: ["./excel-export.component.scss"],
+    templateUrl: "./excel-export.component.html"
+})
+export class ExcelExportComponent {
+    constructor(private excelExportService: IgxExcelExporterService) {
+        this.excelExportService.onExportEnded.subscribe();
+    }
+}
+@NgModule({
+    declarations: [ExcelExportComponent],
+    exports: [ExcelExportComponent],
+    imports: [],
+    providers: [IgxExcelExporterService]
+});
+`);
+
+        const tree = await runner
+            .runSchematicAsync('migration-19', {}, appTree)
+            .toPromise();
+
+        const expectedContent =
+`import { Component } from '@angular/core';
+import { IgxExcelExporterService } from "igniteui-angular";
+
+@Component({
+    selector: "app-excel-export",
+    styleUrls: ["./excel-export.component.scss"],
+    templateUrl: "./excel-export.component.html"
+})
+export class ExcelExportComponent {
+    constructor(private excelExportService: IgxExcelExporterService) {
+        this.excelExportService.exportEnded.subscribe();
+    }
+}
+@NgModule({
+    declarations: [ExcelExportComponent],
+    exports: [ExcelExportComponent],
+    imports: [],
+    providers: [IgxExcelExporterService]
+});
+`;
+
+        expect(
+            tree.readContent(
+                '/testSrc/appPrefix/component/excel-export.component.ts'
+            )
+        ).toEqual(expectedContent);
+    });
+
+    it('should update CSV exporter onExportEnded event name to exportEnded', async () => {
+        pending('set up tests for migrations through lang service');
+        appTree.create(
+            '/testSrc/appPrefix/component/csv-export.component.ts',
+`import { Component } from '@angular/core';
+import { IgxCsvExporterService } from "igniteui-angular";
+
+@Component({
+    selector: "app-csv-export",
+    styleUrls: ["./csv-export.component.scss"],
+    templateUrl: "./csv-export.component.html"
+})
+export class CsvExportComponent {
+    constructor(private csvExportService: IgxCsvExporterService) {
+        this.csvExportService.onExportEnded.subscribe();
+    }
+}
+@NgModule({
+    declarations: [CsvExportComponent],
+    exports: [CsvExportComponent],
+    imports: [],
+    providers: [IgxCsvExporterService]
+});
+`);
+
+        const tree = await runner
+            .runSchematicAsync('migration-19', {}, appTree)
+            .toPromise();
+
+        const expectedContent =
+`import { Component } from '@angular/core';
+import { IgxCsvExporterService } from "igniteui-angular";
+
+@Component({
+    selector: "app-csv-export",
+    styleUrls: ["./csv-export.component.scss"],
+    templateUrl: "./csv-export.component.html"
+})
+export class CsvExportComponent {
+    constructor(private csvExportService: IgxCsvExporterService) {
+        this.csvExportService.exportEnded.subscribe();
+    }
+}
+@NgModule({
+    declarations: [CsvExportComponent],
+    exports: [CsvExportComponent],
+    imports: [],
+    providers: [IgxCsvExporterService]
+});
+`;
+        expect(
+            tree.readContent(
+                '/testSrc/appPrefix/component/csv-export.component.ts'
+            )
+        ).toEqual(expectedContent);
+    });
+
     it('should replace onSelect and onUnselect with selected and deselected in igx-buttongroup', async () => {
         appTree.create(
             `/testSrc/appPrefix/component/buttongroup.component.html`,
@@ -308,6 +521,131 @@ export class IconTestComponent {
                 (hidden)="someHandler($event)"
             >
             </igx-toast>`
+        );
+    });
+
+    it('should replace IgxTooltipTargetDirective event names onTooltipShow and onTooltipHide with tooltipShow and tooltipHide',
+        async () => {
+            appTree.create(
+            `/testSrc/appPrefix/component/tooltip.component.html`,
+    `<button [igxTooltipTarget]="tooltipRef" (onTooltipShow)="showing($event)" (onTooltipHide)="hiding($event)">
+        Hover me
+    </button>
+    <div igxTooltip #tooltipRef="tooltip">
+        Hello, I am a tooltip!
+    </div>`
+        );
+
+        const tree = await runner
+            .runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/tooltip.component.html')
+        ).toEqual(
+    `<button [igxTooltipTarget]="tooltipRef" (tooltipShow)="showing($event)" (tooltipHide)="hiding($event)">
+        Hover me
+    </button>
+    <div igxTooltip #tooltipRef="tooltip">
+        Hello, I am a tooltip!
+    </div>`);
+    });
+
+    it('should replace outputs with selected in igx-calendar', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/calendar.component.html`,
+`<igx-calendar
+    (onSelection)="someHandler($event)"
+    (onViewChanging)="someHandler($event)"
+    (onDateSelection)="someHandler($event)"
+    (onYearSelection)="someHandler($event)"
+    (onMonthSelection)="someHandler($event)"
+></igx-calendar>`
+        );
+
+        const tree = await runner.runSchematicAsync(migrationName, {}, appTree).toPromise();
+
+        expect(tree.readContent('/testSrc/appPrefix/component/calendar.component.html')).toEqual(
+`<igx-calendar
+    (selected)="someHandler($event)"
+    (viewChanging)="someHandler($event)"
+    (dateSelection)="someHandler($event)"
+    (yearSelection)="someHandler($event)"
+    (monthSelection)="someHandler($event)"
+></igx-calendar>`
+        );
+    });
+
+    it('should replace onSelection and onYearSelection with selected in igx-years-view', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/yearsview.component.html`,
+`<igx-years-view
+    (onSelection)="changeYear($event)"
+    (onYearSelection)="changeYear($event)"
+></igx-years-view>`
+        );
+
+        const tree = await runner.runSchematicAsync(migrationName, {}, appTree).toPromise();
+
+        expect(tree.readContent('/testSrc/appPrefix/component/yearsview.component.html')).toEqual(
+`<igx-years-view
+    (selected)="changeYear($event)"
+    (yearSelection)="changeYear($event)"
+></igx-years-view>`
+        );
+    });
+
+    it('should replace onDateSelection and onViewChanging with selected in igx-days-view', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/daysview.component.html`,
+`<igx-days-view
+    (onDateSelection)="someHandler($event)"
+    (onViewChanging)="someHandler($event)"
+></igx-days-view>`
+        );
+
+        const tree = await runner.runSchematicAsync(migrationName, {}, appTree).toPromise();
+
+        expect(tree.readContent('/testSrc/appPrefix/component/daysview.component.html')).toEqual(
+`<igx-days-view
+    (dateSelection)="someHandler($event)"
+    (viewChanging)="someHandler($event)"
+></igx-days-view>`
+        );
+    });
+
+    it('should replace onSelection and onMonthSelection with selected in igx-months-view', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/monthsview.component.html`,
+`<igx-months-view
+    (onSelection)="someHandler($event)"
+    (onMonthSelection)="someHandler($event)"
+></igx-months-view>`
+        );
+
+        const tree = await runner.runSchematicAsync(migrationName, {}, appTree).toPromise();
+
+        expect(tree.readContent('/testSrc/appPrefix/component/monthsview.component.html')).toEqual(
+`<igx-months-view
+    (selected)="someHandler($event)"
+    (monthSelection)="someHandler($event)"
+></igx-months-view>`
+        );
+    });
+
+    it('should replace onSelection and onMonthSelection with selected in igx-month-picker', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/monthpicker.component.html`,
+`<igx-month-picker
+    (onSelection)="someHandler($event)"
+></igx-month-picker>`
+        );
+
+        const tree = await runner.runSchematicAsync(migrationName, {}, appTree).toPromise();
+
+        expect(tree.readContent('/testSrc/appPrefix/component/monthpicker.component.html')).toEqual(
+`<igx-month-picker
+    (selected)="someHandler($event)"
+></igx-month-picker>`
         );
     });
 });