+++ projects/igniteui-angular/src/lib/grids/filtering/excel-style/grid.excel-style-filtering.component.ts
@@ -79,7 +79,7 @@ export class IgxGridExcelStyleFilteringComponent implements OnDestroy {
      * @hidden @internal
      */
     @HostBinding('class.igx-excel-filter')
-    className = 'igx-excel-filter';
+    public defaultClass = true;
 
     /**
      * @hidden @internal
@@ -254,7 +254,7 @@ export class IgxGridExcelStyleFilteringComponent implements OnDestroy {
      * Gets the minimum height.
      */
     @Input()
-    get minHeight(): string {
+    public get minHeight(): string {
         if (this._minHeight || this._minHeight === 0) {
             return this._minHeight;
         }
@@ -273,7 +273,7 @@ export class IgxGridExcelStyleFilteringComponent implements OnDestroy {
     /**
      * Sets the minimum height.
      */
-    set minHeight(value: string) {
+    public set minHeight(value: string) {
         this._minHeight = value;
     }
 
@@ -298,7 +298,7 @@ export class IgxGridExcelStyleFilteringComponent implements OnDestroy {
      */
     @Input()
     @HostBinding('style.max-height')
-    get maxHeight(): string {
+    public get maxHeight(): string {
         if (this._maxHeight) {
             return this._maxHeight;
         }
@@ -317,21 +317,21 @@ export class IgxGridExcelStyleFilteringComponent implements OnDestroy {
     /**
      * Sets the maximum height.
      */
-    set maxHeight(value: string) {
+    public set maxHeight(value: string) {
         this._maxHeight = value;
     }
 
     /**
      * @hidden @internal
      */
-    get grid(): IgxGridBaseDirective {
+    public get grid(): IgxGridBaseDirective {
         return this.gridAPI?.grid ?? this.column?.grid;
     }
 
     /**
      * @hidden @internal
      */
-    get displayDensity() {
+    public get displayDensity() {
         return this.grid?.displayDensity;
     }
 
@@ -343,7 +343,7 @@ export class IgxGridExcelStyleFilteringComponent implements OnDestroy {
     /**
      * @hidden @internal
      */
-    ngOnDestroy(): void {
+    public ngOnDestroy(): void {
         this.destroy$.next(true);
         this.destroy$.complete();
     }
@@ -405,8 +405,7 @@ export class IgxGridExcelStyleFilteringComponent implements OnDestroy {
      * @hidden @internal
      */
     public onHideToggle() {
-        this.column.hidden = !this.column.hidden;
-        this.grid.onColumnVisibilityChanged.emit({ column: this.column, newValue: this.column.hidden });
+        this.column.toggleVisibility();
         this.closeDropdown();
     }