+++ projects/igniteui-angular/src/lib/grids/grid-base.directive.ts
@@ -37,7 +37,7 @@ import { cloneArray, flatten, mergeObjects, isIE, compareMaps, resolveNestedPath
 import { DataType } from '../data-operations/data-util';
 import { FilteringLogic, IFilteringExpression } from '../data-operations/filtering-expression.interface';
 import { IGroupByRecord } from '../data-operations/groupby-record.interface';
-import { ISortingExpression } from '../data-operations/sorting-expression.interface';
+import { ISortingExpression, SortingDirection } from '../data-operations/sorting-expression.interface';
 import { IgxGridForOfDirective } from '../directives/for-of/for_of.directive';
 import { IgxTextHighlightDirective } from '../directives/text-highlight/text-highlight.directive';
 import {
@@ -54,7 +54,6 @@ import {
 } from '../services/public_api';
 import { GridBaseAPIService } from './api.service';
 import { IgxGridCellComponent } from './cell.component';
-import { IColumnVisibilityChangedEventArgs } from './column-actions/column-hiding.directive';
 import { ISummaryExpression } from './summaries/grid-summary';
 import { RowEditPositionStrategy, IPinningConfig } from './grid.common';
 import { IgxGridToolbarComponent } from './toolbar/grid-toolbar.component';
@@ -118,7 +117,6 @@ import {
     IRowSelectionEventArgs,
     IPinColumnEventArgs,
     IGridEditEventArgs,
-    IPageEventArgs,
     IRowDataEventArgs,
     IColumnResizeEventArgs,
     IColumnMovingStartEventArgs,
@@ -136,7 +134,12 @@ import {
     IPinRowEventArgs,
     IGridScrollEventArgs,
     IGridEditDoneEventArgs,
-    IActiveNodeChangeEventArgs
+    IActiveNodeChangeEventArgs,
+    ISortingEventArgs,
+    IFilteringEventArgs,
+    IColumnVisibilityChangedEventArgs,
+    IColumnVisibilityChangingEventArgs,
+    IPinColumnCancellableEventArgs
 } from './common/events';
 import { IgxAdvancedFilteringDialogComponent } from './filtering/advanced-filtering/advanced-filtering-dialog.component';
 import { GridType } from './common/grid.interface';
@@ -287,6 +290,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     public onScroll = new EventEmitter<IGridScrollEventArgs>();
 
     /**
+     * @deprecated Use `IgxPaginator` corresponding output instead.
      * Emitted after the current page is changed.
      *
      * @example
@@ -299,10 +303,12 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      * }
      * ```
      */
+    @DeprecateProperty('Use the corresponding output exposed by the `igx-paginator`.')
     @Output()
     public pageChange = new EventEmitter<number>();
 
     /**
+     * @deprecated Use `IgxPaginator` corresponding output instead.
      * Emitted when `perPage` property value of the grid is changed.
      *
      * @example
@@ -315,6 +321,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      * }
      * ```
      */
+    @DeprecateProperty('Use the corresponding output exposed by the `igx-paginator`.')
     @Output()
     public perPageChange = new EventEmitter<number>();
 
@@ -447,7 +454,24 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     public onColumnSelectionChange = new EventEmitter<IColumnSelectionEventArgs>();
 
     /**
-     * Emitted when `IgxColumnComponent` is pinned.
+     * Emitted before `IgxColumnComponent` is pinned.
+     *
+     * @remarks
+     * The index at which to insert the column may be changed through the `insertAtIndex` property.
+     * @example
+     * ```typescript
+     * public columnPinning(event) {
+     *     if (event.column.field === "Name") {
+     *       event.insertAtIndex = 0;
+     *     }
+     * }
+     * ```
+     */
+    @Output()
+    public onColumnPinning = new EventEmitter<IPinColumnCancellableEventArgs>();
+
+    /**
+     * Emitted after `IgxColumnComponent` is pinned.
      *
      * @remarks
      * The index that the column is inserted at may be changed through the `insertAtIndex` property.
@@ -461,7 +485,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      * ```
      */
     @Output()
-    public onColumnPinning = new EventEmitter<IPinColumnEventArgs>();
+    public columnPinned = new EventEmitter<IPinColumnEventArgs>();
 
     /**
      * Emitted when cell enters edit mode.
@@ -596,7 +620,20 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     public onColumnInit = new EventEmitter<IgxColumnComponent>();
 
     /**
-     * Emitted when sorting is performed through the UI.
+     * Emitted before sorting expressions are applied.
+     *
+     * @remarks
+     * Returns an `ISortingEventArgs` object. `sortingExpressions` key holds the sorting expressions.
+     * @example
+     * ```html
+     * <igx-grid #grid [data]="localData" [autoGenerate]="true" (onSorting)="sorting($event)"></igx-grid>
+     * ```
+     */
+    @Output()
+    public sorting = new EventEmitter<ISortingEventArgs>();
+
+    /**
+     * Emitted after sorting is completed.
      *
      * @remarks
      * Returns the sorting expression.
@@ -609,7 +646,20 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     public onSortingDone = new EventEmitter<ISortingExpression | Array<ISortingExpression>>();
 
     /**
-     * Emitted when filtering is performed through the UI.
+     * Emitted before filtering expressions are applied.
+     *
+     * @remarks
+     * Returns an `IFilteringEventArgs` object. `filteringExpressions` key holds the filtering expressions for the column.
+     * @example
+     * ```html
+     * <igx-grid #grid [data]="localData" [height]="'305px'" [autoGenerate]="true" (filtering)="filtering($event)"></igx-grid>
+     * ```
+     */
+    @Output()
+    public filtering = new EventEmitter<IFilteringEventArgs>();
+
+    /**
+     * Emitted after filtering is performed through the UI.
      *
      * @remarks
      * Returns the filtering expressions tree of the column for which filtering was performed.
@@ -620,19 +670,6 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      */
     @Output()
     public onFilteringDone = new EventEmitter<IFilteringExpressionsTree>();
-
-    /**
-     * Emitted when paging is performed.
-     *
-     * @remarks
-     * Returns an object consisting of the previous and next pages.
-     * @example
-     * ```html
-     * <igx-grid #grid [data]="localData" [height]="'305px'" [autoGenerate]="true" (onPagingDone)="pagingDone($event)"></igx-grid>
-     * ```
-     */
-    @Output()
-    public onPagingDone = new EventEmitter<IPageEventArgs>();
 
     /**
      * Emitted when a row added through the API.
@@ -661,7 +698,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     public onRowDeleted = new EventEmitter<IRowDataEventArgs>();
 
     /**
-     * Emitted when column is resized.
+     * Emitted after column is resized.
      *
      * @remarks
      * Returns the `IgxColumnComponent` object's old and new width.
@@ -699,10 +736,23 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     public onDoubleClick = new EventEmitter<IGridCellEventArgs>();
 
     /**
-     * Emitted when column visibility is changed.
+     * Emitted before column visibility is changed.
      *
      * @remarks
      * Args: { column: any, newValue: boolean }
+     * @example
+     * ```html
+     * <igx-grid [columnHiding]="true" [showToolbar]="true" (columnVisibilityChanging)="visibilityChanging($event)"></igx-grid>
+     * ```
+     */
+   @Output()
+   public columnVisibilityChanging = new EventEmitter<IColumnVisibilityChangingEventArgs>();
+
+    /**
+     * Emitted after column visibility is changed.
+     *
+     * @remarks
+     * Args: { column: IgxColumnComponent, newValue: boolean }
      * @example
      * ```html
      * <igx-grid [columnHiding]="true" [showToolbar]="true" (onColumnVisibilityChanged)="visibilityChanged($event)"></igx-grid>
@@ -1170,9 +1220,9 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     @ViewChild('defaultCollapsedTemplate', { read: TemplateRef, static: true })
     protected defaultCollapsedTemplate: TemplateRef<any>;
 
-     /**
-      * @hidden @internal
-      */
+    /**
+     * @hidden @internal
+     */
     @ViewChild('defaultESFHeaderIcon', { read: TemplateRef, static: true })
     protected defaultESFHeaderIconTemplate: TemplateRef<any>;
 
@@ -1371,6 +1421,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     }
 
     /**
+     * @deprecated Use `IgxPaginator` corresponding method instead.
      * Gets/Sets the current page index.
      *
      * @example
@@ -1380,6 +1431,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      * @remarks
      * Supports two-way binding.
      */
+    @DeprecateProperty('Use the corresponding method exposed by the `igx-paginator`.')
     @Input()
     public get page(): number {
         return this._page;
@@ -1390,7 +1442,6 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
             return;
         }
         this.selectionService.clear(true);
-        this.onPagingDone.emit({ previous: this._page, current: val });
         this._page = val;
         this.pageChange.emit(this._page);
         this.navigateTo(0);
@@ -1398,6 +1449,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     }
 
     /**
+     * @deprecated Use `IgxPaginator` corresponding method instead.
      * Gets/Sets the number of visible items per page.
      *
      * @remarks
@@ -1407,6 +1459,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      * <igx-grid #grid [data]="Data" [paging]="true" [(perPage)]="model.perPage" [autoGenerate]="true"></igx-grid>
      * ```
      */
+    @DeprecateProperty('Use the corresponding method exposed by the `igx-paginator`.')
     @Input()
     public get perPage(): number {
         return this._perPage;
@@ -1419,7 +1472,9 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
         this.selectionService.clear(true);
         this._perPage = val;
         this.perPageChange.emit(this._perPage);
-        this.page = 0;
+        if (this.totalPages !== 0 && this._page >= this.totalPages) {
+            this.page = this.totalPages - 1;
+        }
         this.endEdit(true);
         this.notifyChanges();
     }
@@ -2848,6 +2903,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     private _showSummaryOnCollapse = false;
     private _cellSelectionMode: GridSelectionMode = GridSelectionMode.multiple;
     private _rowSelectionMode: GridSelectionMode = GridSelectionMode.none;
+    private _selectRowOnClick = true;
     private _columnSelectionMode: GridSelectionMode = GridSelectionMode.none;
 
     private lastAddedRowIndex;
@@ -3006,6 +3062,57 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
         this.datePipe = new DatePipe(this.locale);
         this.decimalPipe = new DecimalPipe(this.locale);
         this.cdr.detach();
+    }
+
+    /**
+     * @deprecated Use `IgxPaginator` corresponding method instead.
+     * Goes to the desired page index.
+     *
+     * @example
+     * ```typescript
+     * this.grid1.paginate(1);
+     * ```
+     * @param val
+     */
+    @DeprecateProperty('Use the corresponding method exposed by the `igx-paginator`.')
+    public paginate(val: number): void {
+        if (val < 0 || val > this.totalPages - 1) {
+            return;
+        }
+
+        this.page = val;
+    }
+
+    /**
+     * @deprecated Use `IgxPaginator` corresponding method instead.
+     * Goes to the next page, if the grid is not already at the last page.
+     *
+     * @example
+     * ```typescript
+     * this.grid1.nextPage();
+     * ```
+     */
+    @DeprecateProperty('Use the corresponding method exposed by the `igx-paginator`.')
+    public nextPage(): void {
+        if (!this.isLastPage) {
+            this.page += 1;
+        }
+    }
+
+    /**
+     * @deprecated Use `IgxPaginator` corresponding method instead.
+     * Goes to the previous page, if the grid is not already at the first page.
+     *
+     * @example
+     * ```typescript
+     * this.grid1.previousPage();
+     * ```
+     */
+    @DeprecateProperty('Use the corresponding method exposed by the `igx-paginator`.')
+    public previousPage(): void {
+        if (!this.isFirstPage) {
+            this.page -= 1;
+        }
     }
 
     /**
@@ -3226,11 +3333,6 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
                 this.zone.run(() => {
                     this.notifyChanges(true);
                 });
-            });
-
-        this.onPagingDone.pipe(destructor).subscribe(() => {
-            this.endEdit(true);
-            this.selectionService.clear(true);
         });
 
         this.onColumnMoving.pipe(destructor).subscribe(() => this.endEdit(true));
@@ -3306,6 +3408,12 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
             this.summaryService.summaryHeight = 0;
             this.notifyChanges(true);
         });
+    }
+
+    /** @hidden @internal */
+    public _pagingDone() {
+        this.endEdit(true);
+        this.selectionService.clear(true);
     }
 
     /**
@@ -3594,9 +3702,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
         if (!col) {
             return;
         }
-
-        col.hidden = args.newValue;
-        this.onColumnVisibilityChanged.emit(args);
+        col.toggleVisibility(args.newValue);
     }
 
     /**
@@ -4041,6 +4147,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     }
 
     /**
+     * @deprecated Use `IgxPaginator` corresponding method instead.
      * Gets the total number of pages.
      *
      * @example
@@ -4048,6 +4155,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      * const totalPages = this.grid.totalPages;
      * ```
      */
+    @DeprecateProperty('Use the corresponding method exposed by the `igx-paginator`.')
     public get totalPages(): number {
         if (this.pagingState) {
             return this.pagingState.metadata.countPages;
@@ -4056,6 +4164,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     }
 
     /**
+     * @deprecated Use `IgxPaginator` corresponding method instead.
      * Gets if the current page is the first page.
      *
      * @example
@@ -4063,36 +4172,9 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      * const firstPage = this.grid.isFirstPage;
      * ```
      */
+    @DeprecateProperty('Use the corresponding method exposed by the `igx-paginator`.')
     public get isFirstPage(): boolean {
         return this.page === 0;
-    }
-
-    /**
-     * Goes to the next page, if the grid is not already at the last page.
-     *
-     * @example
-     * ```typescript
-     * this.grid1.nextPage();
-     * ```
-     */
-    public nextPage(): void {
-        if (!this.isLastPage) {
-            this.page += 1;
-        }
-    }
-
-    /**
-     * Goes to the previous page, if the grid is not already at the first page.
-     *
-     * @example
-     * ```typescript
-     * this.grid1.previousPage();
-     * ```
-     */
-    public previousPage(): void {
-        if (!this.isFirstPage) {
-            this.page -= 1;
-        }
     }
 
     /**
@@ -4119,6 +4201,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     }
 
     /**
+     * @deprecated Use `IgxPaginator` corresponding method instead.
      * Returns if the current page is the last page.
      *
      * @example
@@ -4126,6 +4209,7 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      * const lastPage = this.grid.isLastPage;
      * ```
      */
+    @DeprecateProperty('Use the corresponding method exposed by the `igx-paginator`.')
     public get isLastPage(): boolean {
         return this.page + 1 >= this.totalPages;
     }
@@ -4197,41 +4281,31 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
             this._moveChildColumns(column.parent, column, target, pos);
         }
 
+        let columnPinStateChanged;
+        // pinning and unpinning will work correctly even without passing index
+        // but is easier to calclulate the index here, and later use it in the pinning event args
         if (target.pinned && !column.pinned) {
-            column.pin();
+            const pinnedIndex = this._pinnedColumns.indexOf(target);
+            const index = pos === DropPosition.AfterDropTarget ? pinnedIndex + 1 : pinnedIndex;
+            columnPinStateChanged = column.pin(index);
         }
 
         if (!target.pinned && column.pinned) {
-            column.unpin();
+            const unpinnedIndex = this._unpinnedColumns.indexOf(target);
+            const index = pos === DropPosition.AfterDropTarget ? unpinnedIndex + 1 : unpinnedIndex;
+            columnPinStateChanged = column.unpin(index);
         }
 
-        if (target.pinned && column.pinned) {
+        if (target.pinned && column.pinned && !columnPinStateChanged) {
             this._reorderColumns(column, target, pos, this._pinnedColumns);
         }
 
-        if (!target.pinned && !column.pinned) {
+        if (!target.pinned && !column.pinned && !columnPinStateChanged) {
             this._reorderColumns(column, target, pos, this._unpinnedColumns);
         }
 
         this._moveColumns(column, target, pos);
         this._columnsReordered(column, target);
-    }
-
-    /**
-     * Goes to the desired page index.
-     *
-     * @example
-     * ```typescript
-     * this.grid1.paginate(1);
-     * ```
-     * @param val
-     */
-    public paginate(val: number): void {
-        if (val < 0 || val > this.totalPages - 1) {
-            return;
-        }
-
-        this.page = val;
     }
 
     /**
@@ -4445,9 +4519,30 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
      * ```
      */
     public sort(expression: ISortingExpression | Array<ISortingExpression>): void {
+        const sortingState = cloneArray(this.sortingExpressions);
+
+        if (expression instanceof Array) {
+            for (const each of expression) {
+                if (each.dir === SortingDirection.None) {
+                    this.gridAPI.remove_grouping_expression(each.fieldName);
+                }
+                this.gridAPI.prepare_sorting_expression([sortingState], each);
+            }
+        } else {
+            if (expression.dir === SortingDirection.None) {
+                this.gridAPI.remove_grouping_expression(expression.fieldName);
+            }
+            this.gridAPI.prepare_sorting_expression([sortingState], expression);
+        }
+
+        const eventArgs: ISortingEventArgs = {owner: this, sortingExpressions: sortingState, cancel: false };
+        this.sorting.emit(eventArgs);
+
+        if (eventArgs.cancel) {
+            return;
+        }
+
         this.endEdit(false);
-
-
         if (expression instanceof Array) {
             this.gridAPI.sort_multiple(expression);
         } else {
@@ -5165,6 +5260,23 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
         return this.isRowPinningToTop ?
             [...this.pinnedDataView, ...this.unpinnedDataView] :
             [...this.unpinnedDataView, ...this.pinnedDataView];
+    }
+
+    /**
+     * Gets/Sets whether clicking over a row should select/deselect it
+     *
+     * @remarks
+     * By default it is set to true
+     * @param enabled: boolean
+     */
+    @WatchChanges()
+    @Input()
+    get selectRowOnClick() {
+        return this._selectRowOnClick;
+    }
+
+    set selectRowOnClick(enabled: boolean) {
+        this._selectRowOnClick = enabled;
     }
 
     /**