+++ src/app/app.module.ts
@@ -141,6 +141,8 @@ import { GridFinJSComponent } from './grid-finjs/grid-finjs.component';
 import { MainComponent } from './grid-finjs/main.component';
 import { ControllerComponent } from './grid-finjs/controllers.component';
 import { CommonModule } from '@angular/common';
+import { GridEventsComponent } from './grid-events/grid-events.component';
+import { PagingPipe } from './grid-events/pagingPipe';
 
 const components = [
     ActionStripSampleComponent,
@@ -172,6 +174,7 @@ const components = [
     ListPanningSampleComponent,
     ListPerformanceSampleComponent,
     MaskSampleComponent,
+    PagingPipe,
     DateTimeEditorSampleComponent,
     NavbarSampleComponent,
     NavdrawerSampleComponent,
@@ -254,6 +257,7 @@ const components = [
     GridFlexSampleComponent,
     GridEsfLoadOnDemandComponent,
     GridFormattingComponent,
+    GridEventsComponent,
     GridFilteringComponent,
     GridFinJSComponent,
     MainComponent,
@@ -300,6 +304,7 @@ const components = [
         HierarchicalRemoteService,
         GridBaseAPIService,
         IgxGridHierarchicalPipe,
+        PagingPipe,
         IgxExcelExporterService,
         IgxIconService,
         IgxCsvExporterService,